requirements-dev.txt
coverage==7.5.3
pytest-cov==5.0.0
pytest==8.2.1
requests-mock==1.12.1
ruff==0.4.6
